fix(mycelix): resolve all clippy warnings in commons and civic clusters
- Collapse nested match arms in bridge integrity validation (commons + civic)
- Replace sort_by with sort_by_key for timestamp/priority sorting
- Use RangeInclusive::contains instead of manual bounds checks
- Remove unnecessary i64 casts in mutualaid-circles and property-registry
- Eliminate redundant identical-branch if-else in property-transfer
- Merge duplicate StoreEntry match arms in justice-evidence integrity
- Replace manual is_multiple_of checks in justice integrity zomes

Both clusters now compile with zero clippy warnings.

+++ mycelix-commons/zomes/mutualaid-circles/integrity/src/lib.rs
@@ -182,7 +182,7 @@ fn validate_credit_line(line: CreditLine) -> ExternResult<ValidateCallbackResult
     }
 
     // Balance cannot exceed credit limit (too negative)
-    if line.balance < -(line.credit_limit as i64) {
+    if line.balance < -line.credit_limit {
         return Ok(ValidateCallbackResult::Invalid(
             "Balance exceeds credit limit".to_string(),
         ));
